Added t-value calculation

+++ views/sampleAScript.js
@@ -41,7 +41,7 @@ document.addEventListener("DOMContentLoaded", () => {
                 for(i=0; i<a; i++) {
                     let value = document.querySelector('.dataValueInputA' + i ).value 
                     totalA += parseInt(value)
-                    squareA += Math.pow(parseInt(value),2)
+                    squareA += Math.pow(parseFloat(value),2)
                 }
                 meanA = totalA / sampleSizeA;
                 localStorage.setItem('totalA', totalA)
